Fix error c2688 when compiling with VS2015
Since C++11, there is such a thing as user-defined literals, which means
it's not possible anymore to append a macro defined literal since it will
treated as a preprocessing token. A simple fix is to leave a whitespace in
order to specify the two literals are separate tokens.

More info :
	* http://stackoverflow.com/questions/31738796/using-macro-with-string-fails-on-vc-2015
	* http://en.cppreference.com/w/cpp/language/user_literal